hosting peculiarities for raleway

+++ server/database/migrations/2024_01_27_215635_create_subscriptions_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->ulid('id')->primary();
             $table->foreignUlid('subscribed_id')->references('id')->on('users');
             $table->foreignUlid('subscriber_id')->references('id')->on('users');
-            $table->unsignedInteger('subscription_amount_paid')->default(0);
+            $table->float('subscription_amount_paid')->default(0);
             $table->timestamp('start_date')->nullable();
             $table->timestamp('expiry_date')->nullable();
             $table->timestamps();
